+++ apps/app/playwright/users.setup.ts
@@ -1,7 +1,7 @@
 import type { APIRequestContext, Browser } from '@playwright/test';
 import { test as setup } from '@playwright/test';
 
-import { activateInvitedUser, inviteUser } from './utils/CreateUser';
+import { activateInvitedUser, inviteUser } from './utils/api';
 import { fillLoginForm, login } from './utils/Login';
 import { FILTER_TEST_USERS, type TestUser } from './utils/test-users';
