+++ apps/app/playwright/utils/api/index.ts
@@ -1,0 +1,7 @@
+// REST API helpers (arrange/teardown). Each takes a Playwright
+// `APIRequestContext`, distinguishing them from the UI-driving helpers in the
+// parent `utils/` directory.
+export * from './CreatePage';
+export * from './CreateUser';
+export * from './CreateUserGroup';
+export * from './SearchIndex';
