+++ apps/app/playwright/utils/index.ts
@@ -1,7 +1,6 @@
+// Public surface of the UI-driving helpers (act on a Playwright `Page`).
+// REST API helpers live behind `utils/api` and are imported from there directly;
+// `test-users` is data imported by path. Neither is re-exported here — the barrel
+// exposes only what barrel callers actually use.
 export * from './CollapseSidebar';
-export * from './CreatePage';
-export * from './CreateUser';
-export * from './CreateUserGroup';
 export * from './Login';
-export * from './SearchIndex';
-export * from './test-users';
